# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GDGOC Hacktoberfest Backend",
    "description": "Event-management API for GDGOC Hacktoberfest open-source contribution tracking.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/contributions": {
      "get": {
        "tags": [
          "Contributions"
        ],
        "summary": "List contributions with moderation filters",
        "description": "List contributions across the platform with filtering by lifecycle status and validation status.\nUseful for maintainers and moderators.",
        "operationId": "get_all_contributions_contributions_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ContributionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by contribution status",
              "title": "Status"
            },
            "description": "Filter by contribution status"
          },
          {
            "name": "validation_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ContributionValidation"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by validation status",
              "title": "Validation Status"
            },
            "description": "Filter by validation status"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContributionResponse"
                  },
                  "title": "Response Get All Contributions Contributions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/contributions/my": {
      "get": {
        "tags": [
          "Contributions"
        ],
        "summary": "Get full contribution timeline for authenticated user",
        "description": "Get the authenticated student's full contribution history and timeline.",
        "operationId": "get_my_timeline_contributions_my_get",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer JWT access token",
              "title": "Authorization"
            },
            "description": "Bearer JWT access token"
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Test/Dev User ID header",
              "title": "X-User-Id"
            },
            "description": "Test/Dev User ID header"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserContributionTimelineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/contributions/{user_id}": {
      "get": {
        "tags": [
          "Contributions"
        ],
        "summary": "Get full contribution timeline for user",
        "description": "Get full contribution history and timeline for a contributor:\nLifecycle progression: claimed \u2192 in_progress \u2192 pr_submitted \u2192 under_review \u2192 changes_requested \u2192 accepted \u2192 merged.",
        "operationId": "get_timeline_contributions__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserContributionTimelineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/contributions/{user_id}/timeline": {
      "get": {
        "tags": [
          "Contributions"
        ],
        "summary": "Get full contribution timeline for user (alias)",
        "description": "Get full contribution history and timeline for a contributor:\nLifecycle progression: claimed \u2192 in_progress \u2192 pr_submitted \u2192 under_review \u2192 changes_requested \u2192 accepted \u2192 merged.",
        "operationId": "get_timeline_contributions__user_id__timeline_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserContributionTimelineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/contributions/{contribution_id}/validation": {
      "patch": {
        "tags": [
          "Contributions"
        ],
        "summary": "Update contribution validation status",
        "description": "Update contribution validation status (valid, pending, rejected, duplicate, invalid).\nTypically restricted to maintainers and organizers.",
        "operationId": "moderate_validation_contributions__contribution_id__validation_patch",
        "parameters": [
          {
            "name": "contribution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Contribution Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer JWT access token",
              "title": "Authorization"
            },
            "description": "Bearer JWT access token"
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Test/Dev User ID header",
              "title": "X-User-Id"
            },
            "description": "Test/Dev User ID header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContributionValidationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContributionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/contributions/{contribution_id}/validate": {
      "patch": {
        "tags": [
          "Contributions"
        ],
        "summary": "Update contribution validation status (alias)",
        "description": "Update contribution validation status (valid, pending, rejected, duplicate, invalid).\nTypically restricted to maintainers and organizers.",
        "operationId": "moderate_validation_contributions__contribution_id__validate_patch",
        "parameters": [
          {
            "name": "contribution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Contribution Id"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bearer JWT access token",
              "title": "Authorization"
            },
            "description": "Bearer JWT access token"
          },
          {
            "name": "X-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Test/Dev User ID header",
              "title": "X-User-Id"
            },
            "description": "Test/Dev User ID header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContributionValidationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContributionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/contributions/{contribution_id}/status": {
      "patch": {
        "tags": [
          "Contributions"
        ],
        "summary": "Transition contribution status",
        "description": "Explicitly advance contribution status according to the valid state machine graph.",
        "operationId": "transition_status_contributions__contribution_id__status_patch",
        "parameters": [
          {
            "name": "contribution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Contribution Id"
            }
          },
          {
            "name": "new_status",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ContributionStatus",
              "description": "Target status in lifecycle state machine"
            },
            "description": "Target status in lifecycle state machine"
          },
          {
            "name": "detail",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional transition note",
              "title": "Detail"
            },
            "description": "Optional transition note"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContributionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClaimUserBrief": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "psit_roll_no": {
            "type": "string",
            "title": "Psit Roll No"
          },
          "github_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Username"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "psit_roll_no"
        ],
        "title": "ClaimUserBrief"
      },
...
      "ContributionPRBrief": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "github_pr_id": {
            "type": "integer",
            "title": "Github Pr Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "github_pr_id",
          "title",
          "status"
        ],
        "title": "ContributionPRBrief"
      },
      "ContributionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "issue_id": {
            "type": "integer",
            "title": "Issue Id"
          },
          "pr_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pr Id"
          },
          "status": {
            "$ref": "#/components/schemas/ContributionStatus"
          },
          "validation_status": {
            "$ref": "#/components/schemas/ContributionValidation"
          },
          "timeline_json": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Timeline Json"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClaimUserBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "issue": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LinkedIssueBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "pull_request": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContributionPRBrief"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "issue_id",
          "status",
          "validation_status",
          "created_at",
          "updated_at"
        ],
        "title": "ContributionResponse"
      },
      "ContributionStatus": {
        "type": "string",
        "enum": [
          "claimed",
          "in_progress",
          "pr_submitted",
          "under_review",
          "changes_requested",
          "accepted",
          "merged"
        ],
        "title": "ContributionStatus"
      },
      "ContributionValidation": {
        "type": "string",
        "enum": [
          "valid",
          "pending",
          "rejected",
          "duplicate",
          "invalid"
        ],
        "title": "ContributionValidation"
      },
      "ContributionValidationUpdate": {
        "properties": {
          "validation_status": {
            "$ref": "#/components/schemas/ContributionValidation"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "validation_status"
        ],
        "title": "ContributionValidationUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LinkedIssueBrief": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "github_issue_id": {
            "type": "integer",
            "title": "Github Issue Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "github_issue_id",
          "title",
          "difficulty",
          "status"
        ],
        "title": "LinkedIssueBrief"
      },
...
      "UserContributionTimelineResponse": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClaimUserBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "total_contributions": {
            "type": "integer",
            "title": "Total Contributions"
          },
          "valid_contributions_count": {
            "type": "integer",
            "title": "Valid Contributions Count"
          },
          "merged_count": {
            "type": "integer",
            "title": "Merged Count"
          },
          "in_progress_count": {
            "type": "integer",
            "title": "In Progress Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ContributionResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "total_contributions",
          "valid_contributions_count",
          "merged_count",
          "in_progress_count",
          "items"
        ],
        "title": "UserContributionTimelineResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}